move errors into separate module and isolate indexing logic
