+++ src/errors.js
@@ -1,0 +1,37 @@
+export class NotImplementedError extends Error {
+  constructor(msg) {
+    super(msg);
+    this.name = 'NotImplementedError';
+  }
+}
+export class NodeNotFoundError extends Error {
+  constructor(msg) {
+    super(msg);
+    this.name = 'NodeNotFoundError';
+  }
+}
+export class IndexError extends Error {
+  constructor(msg) {
+    super(msg);
+    this.name = 'IndexError';
+  }
+}
+export class KeyError extends Error {
+  constructor(msg) {
+    super(msg);
+    this.name = 'KeyError';
+  }
+}
+
+export class ZarrAssertionError extends Error {
+  constructor(msg) {
+    super(msg);
+    this.name = 'ZarrAssertionError';
+  }
+}
+
+export function assert(condition, msg = 'Assertion failed') {
+  if (!condition) {
+    throw new ZarrAssertionError(msg);
+  }
+}
